requirements.txt
streamlit
pillow
folium
pyshp
simplekml
pandas
